+++ src/hexdoc/core/properties.py
@@ -5,7 +5,7 @@ from functools import cached_property
 from pathlib import Path
 from typing import Any, Self, Sequence
 
-from pydantic import Field, PrivateAttr, field_validator
+from pydantic import Field, PrivateAttr, field_validator, model_validator
 from yarl import URL
 
 from hexdoc.model.base import HexdocSettings
@@ -73,6 +73,11 @@ class TemplateProps(StripHiddenModel, validate_assignment=True):
     icon: RelativePath | None = None
     include: PydanticOrderedSet[str]
 
+    render_from: PydanticOrderedSet[str] = Field(None, validate_default=False)
+    """List of modids to include default rendered templates from.
+
+    If not provided, defaults to `self.include`.
+    """
     render: dict[Path, str] = Field(default_factory=dict)
     extend_render: dict[Path, str] = Field(default_factory=dict)
 
@@ -87,14 +92,21 @@ class TemplateProps(StripHiddenModel, validate_assignment=True):
     def override_default_render(self):
         return self._was_render_set
 
-    @field_validator("include", mode="after")
+    @field_validator("include", "render_from", mode="after")
     @classmethod
-    def _resolve_aliases(cls, values: PydanticOrderedSet[str]):
-        for alias, replacement in JINJA_NAMESPACE_ALIASES.items():
-            if alias in values:
-                values.remove(alias)
-                values.add(replacement)
+    def _resolve_aliases(cls, values: PydanticOrderedSet[str] | None):
+        if values:
+            for alias, replacement in JINJA_NAMESPACE_ALIASES.items():
+                if alias in values:
+                    values.remove(alias)
+                    values.add(replacement)
         return values
+
+    @model_validator(mode="after")
+    def _set_default_render_from(self):
+        if self.render_from is None:  # pyright: ignore[reportUnnecessaryComparison]
+            self.render_from = self.include
+        return self
 
 
 # TODO: support item/block override